vulkan::device; re-packaged device into mod
fence got its own module
